requirements.txt
pycryptodome==3.19.0
argon2-cffi==23.1.0
liboqs-python==0.8.0
tk==0.1.0
